screenshots of code
bitbar for some reason is taking netlify url address I paste in and deleting the colon the the url address.
With browser that work like firefox pasted code works first time and with colon in https:
with browsers like edge and chrome they "delete" my pasted colon so manually have to type in colon then works. 
picture prove it. 
LIke to know your thoughts on this,
Sarah.